Move resume's related data to a dedicated directory
